requirements.txt
streamlit
pandas
plotly
google-generativeai
Pillow
matplotlib
openpyxl
xlsxwriter